[Feat] Add Local Getis-Ord G/G* (#7)

+++ src/cpp/weights/weights.h
@@ -27,29 +27,30 @@ double haversine_distance(double lon1, double lat1, double lon2, double lat2, bo
 
 /**
  * @brief Compute the distance weights for a collection of geometries
- * 
+ *
  * @param geoms The geometry collection
  * @param distance_threshold The distance threshold
  * @param is_mile If true, use mile as the distance unit, otherwise use kilometer
  * @return std::vector<std::vector<unsigned int>>  The 2D vector of distance weights
  */
-std::vector<std::vector<unsigned int>> distance_weights(const GeometryCollection& geoms, double distance_threshold,
-                                                        bool is_mile);
+std::vector<std::vector<double>> distance_weights(const GeometryCollection& geoms, double distance_threshold,
+                                                  bool is_mile);
 
 /**
  * @brief Get the distance thresholds, first threshold guarantee that each observation has at least one neighbor, and
  * the second threshold is the maximum distance between two observations
- * 
+ *
  * @param geoms The geometry collection
  * @param is_mile If true, use mile as the distance unit, otherwise use kilometer
- * @return std::vector<double> 
+ * @return std::vector<double>
  */
 std::vector<double> get_distance_thresholds(const GeometryCollection& geoms, bool is_mile);
 
 /**
  * @brief  Compute k-nearest neighbors for a collection of geometries
- * 
- * @param geoms The geometry collection. If the collection is not a point collection, the centroids of the geometries are used.
+ *
+ * @param geoms The geometry collection. If the collection is not a point collection, the centroids of the geometries
+ * are used.
  * @param k The number of nearest neighbors to compute
  * @return std::vector<std::vector<unsigned int>> The 2D vector of k-nearest neighbors
  */
@@ -58,7 +59,7 @@ std::vector<std::vector<unsigned int>> knearest_neighbors(const GeometryCollecti
 /**
  * @brief  Compute contiguity weights for a collection of geometries using the centroids of the geometries.
  * The centroids are used to create a voronoi diagram, which is then used to compute the contiguity weights.
- * 
+ *
  * @param geoms The geometry collection, which could be point/line/polygon collection
  * @param is_queen If true, use queen contiguity, otherwise use rook contiguity
  * @param precision_threshold The precision threshold for comparing coordinates to determine if two points are the same
@@ -71,7 +72,7 @@ std::vector<std::vector<unsigned int>> point_contiguity_weights(const GeometryCo
 
 /**
  * @brief  Compute contiguity weights for a collection of polygons
- * 
+ *
  * @param geoms The polygon collection.
  * @param is_queen If true, use queen contiguity, otherwise use rook contiguity
  * @param precision_threshold  The precision threshold for comparing coordinates to determine if two points are the same
@@ -84,22 +85,22 @@ std::vector<std::vector<unsigned int>> polygon_contiguity_weights(const Geometry
 
 /**
  * @brief Check if two points are equal within a precision threshold
- * 
+ *
  * @param p1 The first point
  * @param p2 The second point
  * @param precision_threshold The precision threshold
- * @return true 
- * @return false 
+ * @return true
+ * @return false
  */
 bool points_equals(const point_type& p1, const point_type& p2, double precision_threshold);
 
 /**
  * @brief Check if two bounding boxes intersect
- * 
+ *
  * @param b1 The first bounding box
  * @param b2 The second bounding box
- * @return true 
- * @return false 
+ * @return true
+ * @return false
  */
 bool bbox_intersects(const box_type& b1, const box_type& b2);
